--- src/domains/cart/hooks/useCart.ts
@@ -1,13 +1,0 @@
-import { useContext } from "react";
-import { CartContext } from "../contexts/CartContext";
-
-const useCart = () => {
-  const context = useContext(CartContext);
-
-  if (!context) {
-    throw new Error("useCart must be used within CartProvider");
-  }
-  return context;
-};
-
-export default useCart;
